Better implementation for the addition of numerical methods to the tensor object.
