+++ numpy/__init__.py
@@ -1,5 +1,10 @@
+# Extend the Tensor object
+import backpass.numpy.core.extend_tensor
+
 from .core.primitives import *
 from .core.wrap_types import *
+
+
 
 from .random import random
