+++ numpy/core/autograd_diff_funcs.py
@@ -1,7 +1,7 @@
 '''
 The MIT License (MIT)
 
-Copyright (c) 2014 by the President, Fellows of Harvard University and Jahdiel Alvarez
+Copyright (c) 2014 by the President, Fellows of Harvard University
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
